+++ packages/editor/web/test/handlePastedText.spec.js
@@ -1,4 +1,6 @@
-import handlePastedText from '../src/RichContentEditor/handlePastedText';
+import handlePastedText, {
+  convertParsedEditorStateObjectToRawData,
+} from '../src/RichContentEditor/handlePastedText';
 import {
   convertFromRaw,
   convertToRaw,
@@ -9,6 +11,8 @@ import {
   raw,
   expectedRaw,
   emptyRaw,
+  rawContent,
+  expectedRawContent,
   headerWithAlignmentGoogleDocsHTML,
   headerWithAlignmentGoogleDocsExpectedRaw,
   headerWithAlignmentWordHTML,
@@ -16,6 +20,8 @@ import {
   textWithLineSpacingWordHTML,
   textWithLineSpacingWordExpectedRaw,
 } from './TestData/pasted-data';
+
+const allHeaders = ['h2', 'h3', 'h4', 'h5', 'h6'];
 
 describe('Paste text tests', () => {
   it('should paste text on atomic block correctly', () => {
@@ -36,7 +42,9 @@ describe('Paste text tests', () => {
     const pastedEditorState = handlePastedText(
       '',
       headerWithAlignmentGoogleDocsHTML,
-      editorStateWithSelection
+      editorStateWithSelection,
+      undefined,
+      allHeaders
     );
     const pastedRaw = convertToRaw(pastedEditorState.getCurrentContent());
     expect(pastedRaw).toEqual(headerWithAlignmentGoogleDocsExpectedRaw);
@@ -50,7 +58,9 @@ describe('Paste text tests', () => {
     const pastedEditorState = handlePastedText(
       '',
       headerWithAlignmentWordHTML,
-      editorStateWithSelection
+      editorStateWithSelection,
+      undefined,
+      allHeaders
     );
     const pastedRaw = convertToRaw(pastedEditorState.getCurrentContent());
     expect(pastedRaw).toEqual(headerWithAlignmentWordExpectedRaw);
@@ -68,5 +78,9 @@ describe('Paste text tests', () => {
     );
     const pastedRaw = convertToRaw(pastedEditorState.getCurrentContent());
     expect(pastedRaw).toEqual(textWithLineSpacingWordExpectedRaw);
+  });
+
+  it('should convert raw content of accordion (from EditorState JSON object to rawData) correctly', () => {
+    expect(convertParsedEditorStateObjectToRawData(rawContent)).toEqual(expectedRawContent);
   });
 });