+++ packages/plugin-accordion/web/src/toolbar/createToolbar.ts
@@ -1,7 +1,7 @@
 import createInlineButtons from './inline-buttons';
 import createInsertButtons from './insert-buttons';
-import { ACCORDION_TYPE } from '../types';
-import { CreatePluginToolbar, TranslationFunction, PluginConfig } from 'wix-rich-content-common';
+import { ACCORDION_TYPE, AccordionPluginEditorConfig } from '../types';
+import { CreatePluginToolbar, TranslationFunction } from 'wix-rich-content-common';
 
 const createToolbar: CreatePluginToolbar = ({
   t,
@@ -9,7 +9,7 @@ const createToolbar: CreatePluginToolbar = ({
   settings,
 }: {
   t: TranslationFunction;
-  settings: PluginConfig;
+  settings: AccordionPluginEditorConfig;
   isMobile: boolean;
 }) => {
   return {