+++ packages/plugin-headers-markdown/web/src/createHeadersMarkdownPlugin.ts
@@ -1,21 +1,24 @@
 import { createBasePlugin } from 'wix-rich-content-plugin-commons';
-import { HEADERS_MARKDOWN_TYPE as type } from './types';
+import { HEADERS_MARKDOWN_TYPE as type, HeadersMarkdownPluginEditorConfig } from './types';
 import { strategy, component } from './decorator';
-import { CreatePluginFunction, PluginConfig } from 'wix-rich-content-common';
+import { CreatePluginFunction } from 'wix-rich-content-common';
 import { DraftDecorator } from 'draft-js';
 import { DEFAULTS } from './defaults';
 
-export const createHeadersMarkdownDecorator = (settings: PluginConfig): DraftDecorator => {
+export const createHeadersMarkdownDecorator = (
+  config: HeadersMarkdownPluginEditorConfig
+): DraftDecorator => {
+  const { [type]: settings = {} } = config;
   return {
     strategy,
     component: props => component({ ...props, ...settings }),
   };
 };
 
-export const createHeadersMarkdownPlugin: CreatePluginFunction = config => {
+export const createHeadersMarkdownPlugin: CreatePluginFunction<HeadersMarkdownPluginEditorConfig> = config => {
   const { [type]: settings = {} } = config;
   const plugin = {
-    decorators: [createHeadersMarkdownDecorator(settings)],
+    decorators: [createHeadersMarkdownDecorator(config)],
   };
 
   return createBasePlugin({ settings, type, defaultPluginData: DEFAULTS, ...config }, plugin);