+++ packages/plugin-html/web/src/createHtmlPlugin.ts
@@ -1,11 +1,11 @@
 import createToolbar from './toolbar/createToolbar';
 import { createBasePlugin } from 'wix-rich-content-plugin-commons';
 import { Component, defaults } from './HtmlComponent';
-import { HTML_TYPE } from './types';
+import { HTML_TYPE, HtmlPluginEditorConfig } from './types';
 import { CreatePluginFunction, Pubsub } from 'wix-rich-content-common';
 import { ComponentType } from 'react';
 
-const createHtmlPlugin: CreatePluginFunction = config => {
+const createHtmlPlugin: CreatePluginFunction<HtmlPluginEditorConfig> = config => {
   const { helpers, isMobile, t, [HTML_TYPE]: settings = {}, getEditorBounds, ...rest } = config;
 
   const simulateEditClick = ({ e, pubsub }: { e: Event; pubsub: Pubsub }) =>