+++ packages/plugin-table/web/src/createTablePlugin.ts
@@ -1,10 +1,10 @@
 import createToolbar from './toolbar/createToolbar';
 import { Component, DEFAULTS } from './table-component';
-import { TABLE_TYPE } from './types';
+import { TABLE_TYPE, TablePluginEditorConfig } from './types';
 import { createBasePlugin } from 'wix-rich-content-plugin-commons';
 import { CreatePluginFunction } from 'wix-rich-content-common';
 
-const createTablePlugin: CreatePluginFunction = config => {
+const createTablePlugin: CreatePluginFunction<TablePluginEditorConfig> = config => {
   const { helpers, t, [TABLE_TYPE]: settings = {}, isMobile, ...rest } = config;
 
   return createBasePlugin({