+++ CHANGELOG.md
@@ -26,6 +26,8 @@
 ### :house: Internal
 - `editor`
   - [#1778](https://github.com/wix-incubator/rich-content/pull/1778) remove react-dom/server
+- `plugin-commons`
+  - [#1771](https://github.com/wix-incubator/rich-content/pull/1771) remove old & incorrect scss classifiers
 
 </details>
 <hr/>